fixed the shape of dummy array pdeldry in lin_strat_chem.F90
The caller passes in the array in the shape of (pcols,pver) while
the shape of the dummy array in the subroutine is (ncol.pver) and
used as such inside subroutine linv3_strat_chem_solve.

When pcols > ncol, undefined values would be used in the calculations
for do3_linoz_du and do3_linoz_psc_du.

[NBFB]